% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{framework}
\alias{framework}
\title{Internal Dataset}
\format{
## `framework`
A data frame with all columns:
\describe{
  \item{tba}{Ctba}
  ...
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
framework
}
\description{
A subset of data
}
\keyword{datasets}
